requirements.txt
numpy~=1.24.2
opencv-python~=4.7.0.68
tqdm~=4.64.1
pyyaml~=6.0
scikit-image~=0.20.0rc5
tflearn~=0.5.0
pillow~=9.4.0
ipython~=8.3.0
zlib~=1.2.13
future~=0.18.3
pygments~=2.11.2
setuptools~=65.6.3
basicsr~=1.3.4.2
scipy~=1.10.1
torchvision~=0.14.1
lmdb~=1.4.0
requests~=2.28.2
lpips~=0.1.4
imageio~=2.25.1
scikit-learn~=1.2.1
wandb~=0.13.10
lxml~=4.9.1